Refactored Package Infrastructure
Hopefully files talk to each other correctly now.
