+++ rl_lab/envs/ceviche/test_env.py
@@ -1,4 +1,4 @@
-from env import CevicheEnv
+from .env import CevicheEnv
 import torch
 import numpy as np
